# apps/api/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/farms/{farmId}/sensors": {
      "post": {
        "operationId": "SensorsController_create",
        "summary": "Criar um novo sensor para uma plantação",
        "parameters": [
          {
            "name": "farmId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSensorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sensor criado com sucesso"
          },
          "400": {
            "description": "Dados inválidos"
          },
          "404": {
            "description": "Plantação não encontrada"
          }
        },
        "tags": [
          "sensors"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "SensorsController_findAll",
        "summary": "Listar todos os sensores de uma plantação",
        "parameters": [
          {
            "name": "farmId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de sensores retornada com sucesso"
          },
          "404": {
            "description": "Plantação não encontrada"
          }
        },
        "tags": [
          "sensors"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/farms/{farmId}/sensors/{id}": {
      "get": {
        "operationId": "SensorsController_findOne",
        "summary": "Buscar um sensor pelo ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sensor encontrado com sucesso"
          },
          "404": {
            "description": "Sensor não encontrado"
          }
        },
        "tags": [
          "sensors"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "SensorsController_update",
        "summary": "Atualizar completamente um sensor pelo ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSensorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sensor atualizado com sucesso"
          },
          "400": {
            "description": "Dados inválidos"
          },
          "404": {
            "description": "Sensor não encontrado"
          }
        },
        "tags": [
          "sensors"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "SensorsController_partialUpdate",
        "summary": "Atualizar parcialmente um sensor pelo ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSensorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sensor atualizado parcialmente com sucesso"
          },
          "400": {
            "description": "Dados inválidos"
          },
          "404": {
            "description": "Sensor não encontrado"
          }
        },
        "tags": [
          "sensors"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "SensorsController_remove",
        "summary": "Remover um sensor pelo ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Sensor removido com sucesso"
          },
          "404": {
            "description": "Sensor não encontrado"
          }
        },
        "tags": [
          "sensors"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/measurements": {
      "post": {
        "operationId": "MeasurementsController_create",
        "summary": "Registrar uma nova medição",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMeasurementDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Medição registrada com sucesso"
          },
          "400": {
            "description": "Dados inválidos"
          }
        },
        "tags": [
          "measurements"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "MeasurementsController_findAll",
        "summary": "Listar todas as medições, com filtro opcional por sensorId",
        "parameters": [
          {
            "name": "sensorId",
            "required": false,
            "in": "query",
            "description": "Ids dos sensores para filtrar as medições",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de medições retornada com sucesso"
          }
        },
        "tags": [
          "measurements"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/measurements/{id}": {
      "get": {
        "operationId": "MeasurementsController_findOne",
        "summary": "Buscar uma medição pelo Id do sensor",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Medição encontrada com sucesso"
          },
          "404": {
            "description": "Medição não encontrada"
          }
        },
        "tags": [
          "measurements"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/alerts": {
      "get": {
        "operationId": "AlertsController_findAll",
        "summary": "Listar todos os alertas com filtro e paginação",
        "parameters": [
          {
            "name": "severity",
            "required": false,
            "in": "query",
            "description": "Severidade do alerta",
            "schema": {
              "example": "ALTA",
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Tipo do alerta",
            "schema": {
              "example": "temperatura",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de alertas retornada com sucesso"
          },
          "401": {
            "description": "Não autorizado"
          }
        },
        "tags": [
          "alerts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/alerts/{id}": {
      "get": {
        "operationId": "AlertsController_findOne",
        "summary": "Buscar um alerta pelo ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Alerta encontrado com sucesso"
          },
          "404": {
            "description": "Alerta não encontrado"
          }
        },
        "tags": [
          "alerts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "API com Swagger",
    "description": "Documentação automática da API com Swagger",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateSensorDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "Sensor de umidade do solo",
            "description": "Tipo de sensor"
          },
          "status": {
            "type": "string",
            "example": "ativo",
            "description": "Status do sensor"
          }
        },
        "required": [
          "type",
          "status"
        ]
      },
      "UpdateSensorDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "Sensor de umidade do solo",
            "description": "Tipo de sensor"
          },
          "status": {
            "type": "string",
            "example": "ativo",
            "description": "Status do sensor"
          }
        },
        "required": [
          "type",
          "status"
        ]
      },
      "CreateMeasurementDto": {
        "type": "object",
        "properties": {
          "sensorId": {
            "type": "string",
            "example": "id gerado pelo prisma na criação do sensor",
            "description": "sensorId do sensor"
          },
          "temperatura": {
            "type": "number",
            "example": "25.5",
            "description": "Temperatura em graus Celsius"
          },
          "umidade": {
            "type": "number",
            "example": "80.0",
            "description": "Umidade em porcentagem"
          },
          "luminosidade": {
            "type": "number",
            "example": "80.0",
            "description": "Luminosidade em porcentagem"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string",
            "example": "2026-06-04T23:22:04.000Z",
            "description": "Timestamp da medição"
          }
        },
        "required": [
          "sensorId",
          "temperatura",
          "umidade",
          "luminosidade",
          "timestamp"
        ]
      },
...
    }
  }
}